refactor: PH2-T02 extract core models and errors into manja-core
Context:

- Phase: PH2
- Task: PH2-T02 Move shared types and errors into manja-core

Changes:

- Added serde/serde_json/chrono/secrecy dependencies and real module roots
  to manja-core for pure models and API-level errors
- Moved Kite HTTP models from manja into manja-core::models and re-exported
  them via the manja kite::connect::models facade
- Extracted KiteApiError and KiteApiException into manja-core::error and
  updated ManjaError to wrap the core error types
- Wired the manja crate to depend on manja-core while preserving the
  existing public re-exports at the manja crate root
- Introduced a resolve_fixture_path helper for tests and updated market and
  session tests to load fixtures correctly in a workspace layout
- Ensured workspace builds and tests succeed with cargo build --all and
  cargo test -p manja

Notes:

- Tests: cargo build --all; cargo test -p manja
- Follow-up: wire manja-http, manja-ticker, and manja-extras to consume
  manja-core types in PH2-T04–PH2-T07
