+++ manja-core/Cargo.toml
@@ -7,4 +7,7 @@ description = "Core types and shared models for the manja SDK"
 repository = "https://github.com/saxena-dev/manja"
 
 [dependencies]
-
+serde = { version = "1.0.202", features = ["derive"] }
+serde_json = "1.0.117"
+chrono = { version = "0.4.38", features = ["serde"] }
+secrecy = "0.8.0"
