+++ manja-core/src/lib.rs
@@ -1,4 +1,8 @@
 //! Core types and shared models for the `manja` SDK.
+//!
+//! This crate hosts transport-agnostic domain models and error descriptions
+//! that are shared across the `manja` workspace. It deliberately avoids
+//! depending on async runtimes or HTTP/WebSocket clients.
 
-// Placeholder module; contents will be extracted from the main crate in Phase 2.
-
+pub mod models;
+pub mod error;
