+++ manja-core/src/models/session.rs
