+++ manja/src/kite/connect/api/market.rs
@@ -215,11 +215,12 @@ mod tests {
 
     #[test]
     fn test_get_full_quotes_success() {
-        let root =
-            std::env::var("CARGO_MANIFEST_DIR").unwrap_or_else(|_| ".".to_string());
-        let path = std::path::Path::new(&root).join("kiteconnect-mocks/quote.json");
-        let json = std::fs::read_to_string(&path)
-            .unwrap_or_else(|err| panic!("failed to read fixture at {}: {}", path.display(), err));
+        let path = crate::kite::connect::client::test_utils::resolve_fixture_path(
+            "kiteconnect-mocks/quote.json",
+        );
+        let json = std::fs::read_to_string(&path).unwrap_or_else(|err| {
+            panic!("failed to read fixture at {}: {}", path.display(), err)
+        });
         let response: KiteApiResponse<HashMap<String, FullQuote>> =
             serde_json::from_str(&json).unwrap();
         let data = response.data.expect("expected quote data");
@@ -229,11 +230,12 @@ mod tests {
 
     #[test]
     fn test_get_ohlc_quotes_success() {
-        let root =
-            std::env::var("CARGO_MANIFEST_DIR").unwrap_or_else(|_| ".".to_string());
-        let path = std::path::Path::new(&root).join("kiteconnect-mocks/ohlc.json");
-        let json = std::fs::read_to_string(&path)
-            .unwrap_or_else(|err| panic!("failed to read fixture at {}: {}", path.display(), err));
+        let path = crate::kite::connect::client::test_utils::resolve_fixture_path(
+            "kiteconnect-mocks/ohlc.json",
+        );
+        let json = std::fs::read_to_string(&path).unwrap_or_else(|err| {
+            panic!("failed to read fixture at {}: {}", path.display(), err)
+        });
         let response: KiteApiResponse<HashMap<String, OHLCQuote>> =
             serde_json::from_str(&json).unwrap();
         let data = response.data.expect("expected quote data");
@@ -243,11 +245,12 @@ mod tests {
 
     #[test]
     fn test_get_ltp_quotes_success() {
-        let root =
-            std::env::var("CARGO_MANIFEST_DIR").unwrap_or_else(|_| ".".to_string());
-        let path = std::path::Path::new(&root).join("kiteconnect-mocks/ltp.json");
-        let json = std::fs::read_to_string(&path)
-            .unwrap_or_else(|err| panic!("failed to read fixture at {}: {}", path.display(), err));
+        let path = crate::kite::connect::client::test_utils::resolve_fixture_path(
+            "kiteconnect-mocks/ltp.json",
+        );
+        let json = std::fs::read_to_string(&path).unwrap_or_else(|err| {
+            panic!("failed to read fixture at {}: {}", path.display(), err)
+        });
         let response: KiteApiResponse<HashMap<String, LTPQuote>> =
             serde_json::from_str(&json).unwrap();
         let data = response.data.expect("expected quote data");
